Decompose simple page to components
